frota_dds/requirements.txt
# Dependências para o exemplo de monitoramento de frota com Cyclone DDS

# Eclipse Cyclone DDS Python
cyclonedds-nightly>=2025.6.6

# Dependências adicionais para funcionalidades extras
numpy>=1.21.0
matplotlib>=3.5.0
pandas>=1.3.0

# Para desenvolvimento e testes (opcional)
pytest>=6.0.0
pytest-cov>=2.12.0